requirements.txt
faster-whisper
fastapi
uvicorn
python-multipart
aiofiles
PyJWT